recipesnap 2/requirements.txt
flask==3.1.0
google-genai>=1.0.0
gunicorn==23.0.0
